Getting rid of imgui
