+++ CMakeLists.txt
@@ -18,7 +18,6 @@ find_package(GLEW REQUIRED)
 find_package(glm REQUIRED)
 
 include(CMakeUtils/files_and_folders.cmake)
-include(${TOTOGL_VENDOR_DIR}/imgui.cmake)
 
 
 file(GLOB_RECURSE TOTOGL_INCLUDE_FILES CONFIGURE_DEPENDS ${TOTOGL_INCLUDE_DIR}/*.hpp ${TOTOGL_INCLUDE_DIR}/*.tpp)
@@ -33,11 +32,10 @@ target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_23)
 set_target_properties(${PROJECT_NAME} PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${TOTOGL_BINARY_DIR})
 set_target_properties(${PROJECT_NAME} PROPERTIES LINKER_LANGUAGE CXX)
 
-target_link_libraries(${PROJECT_NAME} PUBLIC OpenGL glfw GLU GLEW glm imgui)
+target_link_libraries(${PROJECT_NAME} PUBLIC OpenGL glfw GLU GLEW glm)
 
 target_include_directories(${PROJECT_NAME} PUBLIC ${TOTOGL_INCLUDE_DIR})
 target_include_directories(${PROJECT_NAME} PUBLIC ${TOTOGL_VENDOR_DIR}/includes)
-target_include_directories(${PROJECT_NAME} PUBLIC ${IMGUI_INCLUDE_DIR})
 
 
 add_executable(${PROJECT_NAME}_test ${TOTOGL_TEST_SOURCE_FILES})
