+++ include/TotoGL/CameraControl/OrbitControl.hpp
@@ -1,72 +1,21 @@
 #pragma once
 
 #include <TotoGL/GL.hpp>
-#include <imgui.h>
 
 #include "TotoGL/RenderObject/Camera.hpp"
 #include "TotoGL/Window.hpp"
-
-// So, ImGui prevents me from refactoring this class...
 
 namespace TotoGL {
 
 class OrbitControl {
 public:
-    OrbitControl(float angle_x, float angle_y, float distance)
-        : _alpha(angle_x)
-        , _beta(angle_y)
-        , _distance(distance) { }
+    OrbitControl(float angle_x, float angle_y, float distance);
 
-    void apply(Camera& camera) {
-        camera.position() = _position + //
-            glm::vec3(
-                -glm::cos(_alpha) * glm::cos(_beta) * _distance,
-                -glm::sin(_alpha) * _distance,
-                glm::cos(_alpha) * glm::sin(_beta) * _distance);
-        camera.lookAt(_position, { 0, 1, 0 });
-    }
+    void apply(Camera& camera);
 
-    void rotate(float x, float y) {
-        static constexpr auto HALF_PI = glm::half_pi<float>();
-        static constexpr auto EPSILON = .0001f;
-        _alpha += x;
-        _beta += y;
-        if (_alpha > HALF_PI - EPSILON)
-            _alpha = HALF_PI - EPSILON;
-        if (_alpha < -HALF_PI + EPSILON)
-            _alpha = -HALF_PI + EPSILON;
-        _beta = glm::mod(_beta + glm::pi<float>(), glm::tau<float>()) - glm::pi<float>();
-    }
+    void rotate(float x, float y);
 
-    void bindEvents(Window& window) {
-        using InputEventName::MOUSE_BUTTON;
-        using VectorEventName::CURSOR_POSITION;
-        using VectorEventName::SCROLL;
-        static constexpr auto PI = glm::pi<float>();
-
-        static bool is_holding;
-
-        window.on(MOUSE_BUTTON, [&](const InputEvent& event) {
-            if (event.button == GLFW_MOUSE_BUTTON_1) {
-                is_holding = bool(event.action);
-            }
-        });
-        window.on(CURSOR_POSITION, [&](const VectorEvent& event) {
-            if (ImGui::IsWindowFocused(ImGuiFocusedFlags_AnyWindow)) {
-                is_holding = false;
-                return;
-            }
-            if (!is_holding)
-                return;
-            auto [width, height] = window.size();
-            rotate(-event.dy / height * PI, -event.dx / height * PI);
-        });
-        window.on(SCROLL, [&](const VectorEvent& event) {
-            _distance -= event.dy;
-            if (_distance < glm::epsilon<float>())
-                _distance = glm::epsilon<float>();
-        });
-    }
+    void bindEvents(Window& window, std::function<bool()> focus_stolen = empty_function);
 
     glm::vec3& position() { return _position; }
     glm::vec3 position() const { return _position; }
@@ -83,6 +32,8 @@ private:
     float _alpha { 0 };
     float _beta { 0 };
     float _distance { 0 };
+
+    static constexpr auto empty_function = [] { return false; };
 };
 
 } // namespace TotoGL
