--- include/TotoGL/ImGui.hpp
@@ -1,40 +1,0 @@
-#pragma once
-
-#include <functional>
-#include <imgui.h>
-
-#include <backends/imgui_impl_glfw.h>
-#include <backends/imgui_impl_opengl3.h>
-
-#include "Window.hpp"
-
-namespace TotoGL {
-
-// This feels like temporary code, tbh. Like, it's just a useless abstraction layer...
-
-inline void initImGui(Window& window) {
-    IMGUI_CHECKVERSION();
-    ImGui::CreateContext();
-    ImGuiIO& io = ImGui::GetIO();
-    (void)io;
-
-    ImGui_ImplGlfw_InitForOpenGL(window.glfwWindow(), true);
-    ImGui_ImplOpenGL3_Init("#version 460");
-
-    auto& style = ImGui::GetStyle();
-    style.FrameRounding = 2.;
-    style.WindowRounding = 4.;
-}
-
-inline void renderImGui(std::function<void(void)>&& f) {
-    ImGui_ImplOpenGL3_NewFrame();
-    ImGui_ImplGlfw_NewFrame();
-    ImGui::NewFrame();
-
-    f();
-
-    ImGui::Render();
-    ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
-}
-
-} // namespace TotoGL
